# background/docs/openapi3.json
{
  "components": {
    "schemas": {
      "types.Column": {
        "properties": {
          "exp": {
            "description": "expressions, which default to = when the value is null, have =, !=, \u003e, \u003e=, \u003c, \u003c=, like",
            "type": "string"
          },
          "logic": {
            "description": "logical type, default value is \"and\", support \u0026, and, ||, or",
            "type": "string"
          },
          "name": {
            "description": "column name",
            "type": "string"
          },
          "value": {
            "description": "column value"
          }
        },
        "type": "object"
      },
...
      "types.CreateRolesReply": {
        "properties": {
          "code": {
            "description": "return code",
            "type": "integer"
          },
          "data": {
            "description": "return data",
            "properties": {
              "id": {
                "description": "id",
                "type": "integer"
              }
            },
            "type": "object"
          },
          "msg": {
            "description": "return information description",
            "type": "string"
          }
        },
        "type": "object"
      },
      "types.CreateRolesRequest": {
        "properties": {
          "roleCode": {
            "type": "string"
          },
          "roleDesc": {
            "type": "string"
          },
          "roleName": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "types.DeleteRolesByIDReply": {
        "properties": {
          "code": {
            "description": "return code",
            "type": "integer"
          },
          "data": {
            "description": "return data",
            "type": "object"
          },
          "msg": {
            "description": "return information description",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "types.GetRolesByIDReply": {
        "properties": {
          "code": {
            "description": "return code",
            "type": "integer"
          },
          "data": {
            "description": "return data",
            "properties": {
              "roles": {
                "$ref": "#/components/schemas/types.RolesObjDetail"
              }
            },
            "type": "object"
          },
          "msg": {
            "description": "return information description",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "types.ListRolessReply": {
        "properties": {
          "code": {
            "description": "return code",
            "type": "integer"
          },
          "data": {
            "description": "return data",
            "properties": {
              "roless": {
                "items": {
                  "$ref": "#/components/schemas/types.RolesObjDetail"
                },
                "type": "array"
              }
            },
            "type": "object"
          },
          "msg": {
            "description": "return information description",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "types.Params": {
        "properties": {
          "columns": {
            "description": "query conditions",
            "items": {
              "$ref": "#/components/schemas/types.Column"
            },
            "type": "array"
          },
          "limit": {
            "description": "lines per page",
            "type": "integer"
          },
          "page": {
            "description": "page number, starting from page 0",
            "type": "integer"
          },
          "sort": {
            "description": "sorted fields, multi-column sorting separated by commas",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "types.RolesObjDetail": {
        "properties": {
          "createdAt": {
            "type": "string"
          },
          "id": {
            "description": "convert to uint64 id",
            "type": "integer"
          },
          "roleCode": {
            "type": "string"
          },
          "roleDesc": {
            "type": "string"
          },
          "roleName": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "types.UpdateRolesByIDReply": {
        "properties": {
          "code": {
            "description": "return code",
            "type": "integer"
          },
          "data": {
            "description": "return data",
            "type": "object"
          },
          "msg": {
            "description": "return information description",
            "type": "string"
          }
        },
        "type": "object"
      },
      "types.UpdateRolesByIDRequest": {
        "properties": {
          "id": {
            "description": "uint64 id",
            "type": "integer"
          },
          "roleCode": {
            "type": "string"
          },
          "roleDesc": {
            "type": "string"
          },
          "roleName": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "description": "Type Bearer your-jwt-token to Value",
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "http server api docs",
    "title": "godemo api docs",
    "version": "v1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/roles": {
      "post": {
        "description": "Creates a new roles entity using the provided data in the request body.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/types.CreateRolesRequest"
              }
            }
          },
          "description": "roles information",
          "required": true,
          "x-originalParamName": "data"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/types.CreateRolesReply"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Create a new roles",
        "tags": [
          "roles"
        ]
      }
    },
    "/api/v1/roles/list": {
      "post": {
        "description": "Returns a paginated list of roles based on query filters, including page number and size.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/types.Params"
              }
            }
          },
          "description": "query parameters",
          "required": true,
          "x-originalParamName": "data"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/types.ListRolessReply"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get a paginated list of roless by custom conditions",
        "tags": [
          "roles"
        ]
      }
    },
    "/api/v1/roles/{id}": {
      "delete": {
        "description": "Deletes a existing roles identified by the given id in the path.",
        "parameters": [
          {
            "description": "id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/types.DeleteRolesByIDReply"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Delete a roles by id",
        "tags": [
          "roles"
        ]
      },
      "get": {
        "description": "Gets detailed information of a roles specified by the given id in the path.",
        "parameters": [
          {
            "description": "id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/types.GetRolesByIDReply"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get a roles by id",
        "tags": [
          "roles"
        ]
      },
      "put": {
        "description": "Updates the specified roles by given id in the path, support partial update.",
        "parameters": [
          {
            "description": "id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/types.UpdateRolesByIDRequest"
              }
            }
          },
          "description": "roles information",
          "required": true,
          "x-originalParamName": "data"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/types.UpdateRolesByIDReply"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Update a roles by id",
        "tags": [
          "roles"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:8080/"
    },
    {
      "url": "https://localhost:8080/"
    }
  ]
}